위클리2 미션 초기화
