+++ src/main/java/com/devcourse/user/User.java
@@ -1,0 +1,13 @@
+package com.devcourse.user;
+
+import java.util.UUID;
+
+public class User {
+    private final UUID id;
+    private final String name;
+
+    public User(UUID id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+}
